Created README.md, index.html and styles.css
Added the basic information on the README file.
Started developing the first part of the html code, starting from the top and added a bit of customization to it in style.css.
Downloaded the 4 images that will be used in the page, but haven't referenced then in README.md yet.